// src/features/invoices/types.ts
export enum DueDateOption {
  TODAY = 'TODAY',
  TOMORROW = 'TOMORROW',
  SEVEN_DAYS = 'SEVEN_DAYS',
  FOURTEEN_DAYS = 'FOURTEEN_DAYS',
  THIRTY_DAYS = 'THIRTY_DAYS',
  FORTY_FIVE_DAYS = 'FORTY_FIVE_DAYS',
  SIXTY_DAYS = 'SIXTY_DAYS',
  NINETY_DAYS = 'NINETY_DAYS',
  CUSTOM_DATE = 'CUSTOM_DATE',
}

export enum LateFeeOption {
  NO_LATE_FEE = 'NO_LATE_FEE',
  PERCENTAGE = 'PERCENTAGE',
  FIXED_AMOUNT = 'FIXED_AMOUNT',
}